+++ src/backend/typeDefs/mutations/businessObjectModal/businessObjectModal/index.js
@@ -1,7 +1,9 @@
 const createBusinessObjectModalTypeDefs = require('./create')
 const deleteBusinessObjectModalTypeDefs = require('./delete')
+const updateBusinessObjectModalTypeDefs = require('./update')
 
 module.exports = [
   createBusinessObjectModalTypeDefs,
   deleteBusinessObjectModalTypeDefs,
+  updateBusinessObjectModalTypeDefs,
 ]