+++ src/backend/typeDefs/mutations/businessObjectModal/section/update.js
@@ -1,0 +1,12 @@
+const { gql } = require('apollo-server-express')
+
+const updateBusinessObjectModalSectionTypeDefs = gql`
+  input UpdateBusinessObjectModalSectionInput {
+    label: String!
+    modalId: ID!
+    tagId: ID!
+    sectionId: ID!
+  }
+`
+
+module.exports = updateBusinessObjectModalSectionTypeDefs