+++ src/backend/typeDefs/mutations/businessObjectModal/tag/index.js
@@ -1,5 +1,7 @@
 const createBusinessObjectModalTagTypeDefs = require('./create')
+const updateBusinessObjectModalTagTypeDefs = require('./update')
 
 module.exports = [
   createBusinessObjectModalTagTypeDefs,
+  updateBusinessObjectModalTagTypeDefs,
 ]