+++ CHANGELOG.md
@@ -16,6 +16,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 - `Phoenix`: Built default landing page feature for roles and/or users with backend cascade updates to pulse-dev.users.sitemaps
+- `Orion`: Business Object Modals Read, Create, and Update for Modals, Tabs, Sections, and Fields
 
 ## [2.5.0] - Started: 5-11-20